refactor(storage): relocate compute_statistics/dump/{load_snapshot,apply_plan} (S2 2b)
Moves the "4 backend fns" below the seam, per the carve spec: compute_statistics
+ the dump_json/dump_sqlite family (inspect/{statistics,dump}.rs) and the
SQL-touching consolidation phases load_snapshot/apply_plan (consolidation/mod.rs).
storage/sqlite/schema.rs (still in the facade until the next commit) now calls
them down-dep via the facade's inspect/consolidation re-exports.

The facade's dedup/cluster/global submodules split the same way apply_plan
already forced: their SQL-touching apply_dedup/apply_clusters/apply_global
callees moved alongside it (the exact reason apply_plan itself moved), leaving
the pure compute_dedup/compute_clusters/compute_global in the facade. Each
facade file re-exports its moved half so its own #[cfg(test)]-only
local_dedup/cluster_fusion/global_integration wrapper (composing compute_* +
apply_*) keeps resolving unchanged. Per the maintainer's refined JC#1, the
consolidation/mod.rs test module — including the 3 apply_plan_* concurrency
tests — stays in the facade unmodified, calling the local compute_plan plus
the now-re-exported apply_plan.

Per JC#2, statistics.rs/dump.rs's MemoryEngine-level tests (9 total, plus
their 3 insta snapshots, renamed to the new module path) relocate to a new
#[cfg(test)] mod in engine/inspect.rs, where MemoryEngine::statistics()/
dump_state() are actually defined — they build a full MemoryEngine, which
me-backend-sqlite cannot reach. The raw-Connection tests that drive
compute_statistics/stream_snapshot directly move with the production code.

me-backend-sqlite gains: me-traits (ConsolidationConfig/ConsolidationStats),
tempfile promoted from dev to a real dependency (dump.rs's atomic
write-then-rename is production code here), flate2/zstd behind new
compress-gzip/compress-zstd features (forwarded from the facade's own
features), and libc for the unix O_NOFOLLOW dump-safety path.

Many facade imports that survived the extraction only for #[cfg(test)] code
(get_config/set_config, FactStore, Connection, ConsolidationStats, the
apply_dedup/apply_clusters/apply_global/load_snapshot_capped re-exports
themselves) are now #[cfg(test)]-gated too — `cargo clippy --all-targets`
compiles the plain lib target and the test target as separate passes, so an
import used only by test code is a genuine unused-import error on the former.

Full CI-exact gate green (fmt, clippy --all-features -D warnings, build x2,
test --all-features 1967/0, no anti-#903 regression).
