+++ memory-engine/lib/memory-engine/src/inspect/mod.rs
@@ -16,11 +16,17 @@
 // regression) needs only `read_snapshot` as its untrusted-FILE ingest target; that one
 // function is re-exported through the `#[cfg(fuzzing)] fuzz_seam` in `lib.rs` rather
 // than by widening the whole module. See `fuzz/fuzz_targets/snapshot_restore.rs`.
-pub(crate) mod dump;
+//
+// `dump`/`statistics` relocated to `me-backend-sqlite` (Wave 2 #816 / S2, sub-PR 2b):
+// `compute_statistics`/`dump_json`/… are backend-internal (only `storage::sqlite::schema`
+// calls them), so they moved below the seam with the SQL that produces their data. This
+// re-export preserves `crate::inspect::{dump, statistics}` for that one caller — and for
+// `engine/inspect.rs`'s `#[cfg(test)]` tests, which build a `MemoryEngine` and so cannot
+// move into the backend crate themselves.
+pub(crate) use me_backend_sqlite::inspect::{dump, statistics};
 pub(crate) mod explain;
 pub(crate) mod replay;
 pub(crate) mod restore;
-pub(crate) mod statistics;
 pub mod types;
 
 pub use types::*;
